+++ apps/organizations/admin.py
@@ -75,7 +75,7 @@ class BenefitLevelBenefit(NoRelatedInlineAddEditAdminMixin, ReadOnlyOrgLimitedTa
 
 
 @admin.register(Organization)
-class OrganizationAdmin(RevEngineBaseAdmin, VersionAdmin, ReverseModelAdmin):  # pragma: no cover
+class OrganizationAdmin(RevEngineBaseAdmin, VersionAdmin):  # pragma: no cover
     organization_fieldset = (
         (
             "Organization",
@@ -83,6 +83,7 @@ class OrganizationAdmin(RevEngineBaseAdmin, VersionAdmin, ReverseModelAdmin):  #
                 "fields": (
                     "name",
                     "slug",
+                    "country",
                 )
             },
         ),
@@ -99,12 +100,11 @@ class OrganizationAdmin(RevEngineBaseAdmin, VersionAdmin, ReverseModelAdmin):  #
 
     fieldsets = organization_fieldset
 
-    list_display = ["name", "plan"]
+    list_display = ["name", "plan", "country"]
 
-    list_filter = ["name", "plan", "address__state"]
+    list_filter = ["name", "plan", "country"]
 
     inline_type = "stacked"
-    inline_reverse = [("address", {"fields": ["address1", "address2", "city", "state", "postal_code", "country"]})]
 
     readonly_fields = ["name"]
 
@@ -203,7 +203,6 @@ class RevenueProgramAdmin(RevEngineBaseAdmin, VersionAdmin, ReverseModelAdmin, A
     inline_type = "stacked"
     inline_reverse = [
         ("social_meta", {"fields": ["title", "description", "url", "card"]}),
-        ("address", {"fields": ["address1", "address2", "city", "state", "postal_code"]}),
     ]
     inlines = [RevenueProgramBenefitLevelInline]
