+++ apps/pages/serializers.py
@@ -127,7 +127,7 @@ class DonationPageFullDetailSerializer(serializers.ModelSerializer):
         return obj.revenue_program.payment_provider.get_currency_dict()
 
     def get_organization_country(self, obj):
-        return obj.organization.address.country
+        return obj.organization.country
 
     def get_allow_offer_nyt_comp(self, obj):
         if not self.context.get("live", False):
